Add EmbedReply classes, Colors, and Result tuple
- Embed Replies:
Adds several helper classes which can be used to seemlessly send Embeds
with messages, custom coloring and an emoji indicator to any Discord
Messageable. Initially, added SuccessReply and ErrorReply items which
default to green/red and check/cross emojis.

- Colors:
HexColor extends discord.py's Colour class to add the full array of X11
color names/values.

- Result tuple:
Result is a NamedTuple with success:bool and error:str attributes. This
can be used as a return for "validation" methods which need to verify
the success or failure of some check or series of checks, and then
return an error back to the user.